MySQL基准测试文档添加不同主键类型对主键查询和插入性能的影响

+++ demo-mysql-n-mariadb/demo-order-management-app/common/src/main/java/com/future/demo/entity/OrderModel.java
@@ -2,12 +2,21 @@ package com.future.demo.entity;
 
 import lombok.Data;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Data
 public class OrderModel {
-    private Long id;
+    // long 类型
+    /*private Long id;*/
+    // int 类型
+    /*private Integer id;*/
+    // biginteger 类型
+    private BigInteger id;
+    // uuid string 类型
+    /*private String id;*/
+
     private Long userId;
     private LocalDateTime createTime;
     private Status status;
